requirements.txt
fastapi
uvicorn[standard]
httpx
structlog
sqlmodel
alembic
pydantic-settings
pytest
pytest-cov
pyyaml
flake8
mypy
types-PyYAML
